refactor: PieceTest를 PawnTest와 RookTest로 분리
